refactor: fold node cleanup helpers into canvas_node.go
Keep cleanup query APIs next to the other CanvasNode delete helpers
instead of a separate models file.
